feat: reorg files and funcs
