+++ packages/backend/src/tools/changelog/diff.ts
@@ -1,136 +1,13 @@
-import { assert } from '@l2beat/backend-tools'
-import { ContractParameters, DiscoveryOutput } from '@l2beat/discovery-types'
-import { ChainId, EthereumAddress } from '@lz/libs'
+import { ContractParameters } from '@l2beat/discovery-types'
 import diff from 'deep-diff'
 
-import { groupContracts } from './grouping'
-import {
-  ChangelogEntry,
-  FieldDifference,
-  MilestoneEntry,
-  SmartContractOperation,
-} from './types'
+import { FieldDifference } from './types'
 
-export {
-  changelogEntryToFieldDiff,
-  diffContractValues,
-  fieldDiffToChangelogEntry,
-  getDiscoveryChanges,
-}
-
-function getDiscoveryChanges(
-  previousOutput: DiscoveryOutput,
-  currentOutput: DiscoveryOutput,
-): {
-  properties: {
-    modified: ChangelogEntry[]
-    removed: ChangelogEntry[]
-    added: ChangelogEntry[]
-  }
-  milestones: {
-    added: MilestoneEntry[]
-    removed: MilestoneEntry[]
-  }
-} {
-  const { modified, removed, added } = groupContracts(
-    previousOutput,
-    currentOutput,
-  )
-
-  const removedContractsMileStones = removed.map((c) =>
-    contractToMilestoneEntry(c, {
-      chain: currentOutput.chain,
-      blockNumber: currentOutput.blockNumber,
-      operation: 'CONTRACT_REMOVED',
-    }),
-  )
-
-  const addedContractsMileStones = added.map((c) =>
-    contractToMilestoneEntry(c, {
-      chain: currentOutput.chain,
-      blockNumber: currentOutput.blockNumber,
-      operation: 'CONTRACT_ADDED',
-    }),
-  )
-
-  const removedChangelogEntries = removed.flatMap((contract) => {
-    if (!contract.values) {
-      return []
-    }
-    // Diff values against empty object to get all removed fields
-    const valuesDiff = diffContractValues(contract.values, {})
-
-    return valuesDiff.map((valueDifference) =>
-      fieldDiffToChangelogEntry(valueDifference, {
-        contractName: contract.name,
-        contractAddress: contract.address,
-        blockNumber: currentOutput.blockNumber,
-        chainName: currentOutput.chain,
-      }),
-    )
-  })
-
-  const addedChangelogEntries = added.flatMap((contract) => {
-    if (!contract.values) {
-      return []
-    }
-    // Diff values against empty object to get all added fields
-    const valuesDiff = diffContractValues({}, contract.values)
-
-    return valuesDiff.map((valueDifference) =>
-      fieldDiffToChangelogEntry(valueDifference, {
-        contractName: contract.name,
-        contractAddress: contract.address,
-        blockNumber: currentOutput.blockNumber,
-        chainName: currentOutput.chain,
-      }),
-    )
-  })
-
-  const changelogEntries = modified.flatMap(
-    ([previousContract, currentContract]) => {
-      // FIXME: Double check
-      if (!previousContract.values || !currentContract.values) {
-        return []
-      }
-
-      const fieldDifferences = diffContractValues(
-        previousContract.values,
-        currentContract.values,
-      )
-
-      if (fieldDifferences.length === 0) {
-        // Flat map
-        return []
-      }
-
-      return fieldDifferences.map((valueDifference) =>
-        fieldDiffToChangelogEntry(valueDifference, {
-          contractName: currentContract.name,
-          contractAddress: currentContract.address,
-          blockNumber: currentOutput.blockNumber,
-          chainName: currentOutput.chain,
-        }),
-      )
-    },
-  )
-
-  return {
-    properties: {
-      modified: changelogEntries,
-      removed: removedChangelogEntries,
-      added: addedChangelogEntries,
-    },
-    milestones: {
-      added: addedContractsMileStones,
-      removed: removedContractsMileStones,
-    },
-  }
-}
+export { diffContractValues }
 
 function diffContractValues(
-  firstContractValues: ContractParameters['values'],
-  secondContractValues: ContractParameters['values'],
+  firstContractValues: NonNullable<ContractParameters['values']>,
+  secondContractValues: NonNullable<ContractParameters['values']>,
 ): FieldDifference[] {
   const differences = diff(firstContractValues, secondContractValues)
 
@@ -212,57 +89,4 @@ function diffContractValues(
 
 function pathToKey(path?: string[]): string[] {
   return Array.isArray(path) ? path.map((key) => key.toString()) : []
-}
-
-function fieldDiffToChangelogEntry(
-  fieldDiff: FieldDifference,
-  context: {
-    contractName: string
-    contractAddress: EthereumAddress
-    blockNumber: number
-    chainName: string
-  },
-): ChangelogEntry {
-  assert(fieldDiff.key[0], 'no property path')
-
-  return {
-    targetName: context.contractName,
-    targetAddress: context.contractAddress,
-    chainId: ChainId.fromName(context.chainName),
-    blockNumber: context.blockNumber,
-    modificationType: fieldDiff.modificationType,
-    parameterName: fieldDiff.key[0],
-    parameterPath: fieldDiff.key,
-    previousValue: fieldDiff.previous,
-    currentValue: fieldDiff.current,
-  }
-}
-
-function changelogEntryToFieldDiff(
-  changelogEntry: ChangelogEntry,
-): FieldDifference {
-  return {
-    key: changelogEntry.parameterPath,
-    modificationType: changelogEntry.modificationType,
-    previous: changelogEntry.previousValue,
-    current: changelogEntry.currentValue,
-    // Changelog Entry is dummy fallback, there is no type discrimination
-  } as FieldDifference
-}
-
-function contractToMilestoneEntry(
-  contract: ContractParameters,
-  context: {
-    chain: string
-    blockNumber: number
-    operation: SmartContractOperation
-  },
-): MilestoneEntry {
-  return {
-    targetName: contract.name,
-    targetAddress: contract.address,
-    chainId: ChainId.fromName(context.chain),
-    blockNumber: context.blockNumber,
-    operation: context.operation,
-  }
 }
